refactor(route-card): Pull out journey segment code into selector
