+++ configurations/default/messages.yml
@@ -1,4 +1,8 @@
 Systems:
+  CalculatingAccessibility: Calculating accessibility...
+  SelectStart: Select a location to see accessibility
+  SelectEnd: Select an endpoint to see possible trips
+  Show: show
   AccessTitle: Access to
   BaseTitle: Proposed Transit
   ComparisonTitle: Current Transit
