+++ src/containers/application.js
@@ -5,6 +5,7 @@ import * as actions from '../actions'
 import initializeBrowsochrones from '../actions/browsochrones'
 import Application from '../components/application'
 import selectAccessibilityKeys from '../selectors/accessibility-keys'
+import selectJourneysFromTransitive from '../selectors/journeys-from-transitive'
 import selectPointsOfInterest from '../selectors/points-of-interest'
 import selectShowComparison from '../selectors/show-comparison'
 
@@ -12,6 +13,7 @@ function mapStateToProps (state, ownProps) {
   return {
     ...state,
     accessibilityKeys: selectAccessibilityKeys(state, ownProps),
+    journeys: selectJourneysFromTransitive(state, ownProps),
     pointsOfInterest: selectPointsOfInterest(state, ownProps),
     showComparison: selectShowComparison(state, ownProps)
   }
